More scalable ToolButton
Replaces libinterface's ToolButton, a custom button that can display
a pop-up menu on click (along with a custom image). It didn't scale well
with different themes and font-sizes compared to BButtons― which
ToolButton is supposed to look like to begin with.

A BButton-derived class is used instead, so these buttons look more
in-place among other buttons.